refactor(core): extract answer generation to rag_answer (#468)
Closes #468. Third slice of the rag_core decomposition (after
rag_retrieval PR-H1a/b and rag_verifier PR-J1).

Moves 20 functions from rag_core.py:1784-2145 (~362 LOC) to new
rag_answer.py:
- generate_answer (main entry)
- build_claims / build_comparison_claims / build_extract_claims
- make_claim / claim_target / make_citation
- answer_status / answer_status_reason / answer_query_type /
  answer_summary / answer_verification_reasons
- build_insufficiency
- render_answer_text
- best_sentence / metadata_claim_sentences / metadata_field_requested /
  format_metadata_claim_value / sentence_has_verification_topic /
  select_supporting_evidence

ADR 0003 contract preserved
---------------------------
The dict shape stays in rag_answer (`schema_version: 2` literal,
citation contract). CLAUDE.md prohibition against parallel Pydantic
models continues to apply. No schema bump.

Dependency handling
-------------------
Module-level imports in rag_answer (all leaves):
- korean_lexicon (METADATA_CLAIM_LABELS, METADATA_CLAIM_TOPIC_LABELS,
  METADATA_EVIDENCE_LABELS)
- rag_answer_schema (ANSWER_STATUS_*, ANSWER_SCHEMA_VERSION)
- rag_verifier (specific_topics, verification_topics,
  evidence_text_for_verification, PARTIAL_TOPIC_GROUNDING_REASON)

Late-imports inside specific functions (rag_core symbols with many
non-answer call sites — multi-surface utilities):
- make_citation: normalize_regions, normalize_page_span
- best_sentence: sentence_split, tokenize
- metadata_claim_sentences: ordered_unique
- metadata_field_requested: compact_metadata_text

Re-exports from rag_core
------------------------
All 20 public names re-exported so any caller that imported them
from rag_core (orchestration `_phase_build_answer`, tests, scripts)
keeps working unchanged.

Cross-ref updates
-----------------
- CLAUDE.md: repository map entry for rag_answer.py
- scripts/_governance.py: LOAD_BEARING_PATHS adds rag_answer.py
- .github/pull_request_template.md: load-bearing hint includes rag_answer.py
- ADR 0003 has no rag_core path references — no cross-ref update needed

LOC accounting
--------------
rag_core.py: 3,635 → 3,300 (-335)
rag_answer.py: 0 → 446 (new, with 64-line module docstring + imports)

Verification
------------
Full local pytest 838 passed (1 skipped, 0 failed). Five answer-
touching regression gates stay green:
- tests/test_naive_baseline_ranking_invariance.py (ADR 0001 golden)
- tests/test_retrieval_loop_regression.py
- tests/test_fuzzy_retrieval.py (caught missing compact_metadata_text
  import during debugging)
- tests/test_synthetic_judge.py
- tests/test_langgraph_orchestrator_regression.py
